set reasonable file size limits

+++ src/tools/config_gen.zig
@@ -837,7 +837,7 @@ fn extractSnippetFromSignature(allocator: std.mem.Allocator, signature: []const 
 /// Output example: https://github.com/zigtools/zls/blob/0.11.0/src/data/master.zig
 fn generateVersionDataFile(allocator: std.mem.Allocator, version: []const u8, output_path: []const u8, langref_path: []const u8) !void {
     // const langref_source: []const u8 = @embedFile("langref.html.in");
-    const langref_source = try std.fs.cwd().readFileAlloc(allocator, langref_path, std.math.maxInt(usize));
+    const langref_source = try std.fs.cwd().readFileAlloc(allocator, langref_path, std.math.maxInt(u32));
     defer allocator.free(langref_source);
 
     const builtins = try collectBuiltinData(allocator, version, langref_source);
